some renaming
changed the naming of the file structure
