+++ DungeonsAndDatabases/src/test/java/files/DungeonsAndDatabasesApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package files;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
